requirements.txt
asgiref==3.11.1
Django==4.2.29
mssql-django==1.7
pyodbc==5.3.0
python-dotenv==1.2.1
pytz==2026.1.post1
sqlparse==0.5.5
typing_extensions==4.15.0
